fix(release): share plugin publication authority
